Create num_reverse.py to generate reverse of given number.
This program takes the user input and checks weather given input is numeric value or not. Later it reverse the value of given input and prints the result if given input is not a number the program doesn't crash but rather it displays reverse value of given input with message as given input is not an integer value and try again.