Create Binary Search using C++.
This program takes user input as a size of an array and also the values that are to be inserted in an array. After this user gets the option to search value and then binary search is performed.
If number is found in an array it returns output as element found with its index value else it will print element not found in array.